qcom_defconfig: Enable AUTOBOOT_KEYED support
When debug serial port isn't connected, it is at least reported on RB1
that autoboot gets interrupted. It is probably due to random characters
on the UART RX line when disconnected. Lets try to fix this inconsistent
behaviour via enabling AUTOBOOT_KEYED such that the autoboot gets
interrupted only when the <SPACE> key gets pressed.

Reported-by: Dmitry Baryshkov <dmitry.baryshkov@oss.qualcomm.com>
Signed-off-by: Sumit Garg <sumit.garg@oss.qualcomm.com>